Refactor inputhistory related tests. : (Issue-#29)
