+++ src/test/java/com/github/dakusui/jcunit/examples/recorderreplayer/ReplayerExample.java
@@ -12,6 +12,9 @@ import org.junit.runner.RunWith;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+/**
+ * TODO: This example needs to be re-designed soon.
+ */
 @RunWith(JCUnit.class)
 @GenerateCoveringArrayWith(
     engine = @Generator(
