+++ src/test/java/com/github/dakusui/jcunit/tests/examples/inputhistory/DefaultInputHistoryExampleTest.java
@@ -1,7 +1,6 @@
 package com.github.dakusui.jcunit.tests.examples.inputhistory;
 
 
-import com.github.dakusui.jcunit.examples.inputhistory.DefaultInputHistoryExample;
 import com.github.dakusui.jcunit.testutils.Metatest;
 import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
@@ -11,7 +10,7 @@ import org.junit.runner.RunWith;
 public class DefaultInputHistoryExampleTest {
   public static class AllPassing extends Metatest {
     public AllPassing() {
-      super(DefaultInputHistoryExample.AllPassing.class, 17, 0, 0);
+      super(com.github.dakusui.jcunit.examples.inputhistory.AllPassing.class, 17, 0, 0);
     }
 
     @Test
@@ -21,7 +20,7 @@ public class DefaultInputHistoryExampleTest {
   }
   public static class Failing extends Metatest {
     public Failing() {
-      super(DefaultInputHistoryExample.Failing.class, 17, 13, 0);
+      super(com.github.dakusui.jcunit.examples.inputhistory.Failing.class, 17, 13, 0);
     }
 
     @Test
